+++ mem_utils.h
@@ -4,10 +4,10 @@
 #include "fat.h"
 #include "dir.h"
 
-int to_mem_addr(int cyl, int sect);
+unsigned short to_mem_addr(int cyl, int sect);
 
-int to_cylinder_number(int mem_addr);
+unsigned short to_cylinder_number(unsigned short mem_addr);
 
-int to_sector_number(int mem_addr);
+unsigned short to_sector_number(unsigned short mem_addr);
 
 #endif
